+++ modules/api/api_client.py
@@ -13,30 +13,10 @@ import datetime
 import time
 from concurrent.futures import ThreadPoolExecutor, as_completed
 from typing import Callable, Any, List, Optional, Dict, Union
+from .exceptions import APIRequestError, APIStatusError, APIDataError
 
 DATE_FORMAT = "%Y-%m-%d"
 HOUR_FORMAT = "%H"
-
-class WikimediaAPIError(Exception):
-    """Base exception class for this client."""
-    pass
-
-class APIRequestError(WikimediaAPIError):
-    """Raised for network-level errors (e.g., connection refused, timeout)."""
-    def __init__(self, message, request):
-        super().__init__(message)
-        self.request = request
-
-class APIStatusError(WikimediaAPIError):
-    """Raised for non-successful HTTP status codes (4xx or 5xx)."""
-    def __init__(self, message, request, response):
-        super().__init__(message)
-        self.request = request
-        self.response = response
-
-class APIDataError(WikimediaAPIError):
-    """Raised for errors during data processing (e.g., invalid JSON, corrupt archive)."""
-    pass
 
 class Filter:
     """Represents a simple key-value filter for an API query."""
