+++ modules/api/batch.py
@@ -4,6 +4,7 @@ from project import Project
 from language import Language
 from namespace import Namespace
 from size import Size
+from exceptions import DataModelError
 
 
 class Batch:
@@ -25,24 +26,35 @@ class Batch:
 
     @staticmethod
     def from_json(data: dict) -> 'Batch':
-        return Batch(
-            identifier=data['identifier'],
-            version=data['version'],
-            date_modified=datetime.fromisoformat(data['dateModified']),
-            is_part_of=Project.from_json(data['isPartOf']),
-            in_language=Language.from_json(data['inLanguage']),
-            namespace=Namespace.from_json(data['namespace']),
-            size=Size.from_json(data['size'])
-        )
+        """Safely creates a Batch instance from a dictionary, handling potential errors."""
+        if not isinstance(data, dict):
+            raise DataModelError(f"Expected a dict for Batch data but got {type(data).__name__}")
+        try:
+            date_str = data.get('date_modified')
+            date_obj = datetime.fromisoformat(date_str) if date_str else None
+            
+            return Batch(
+                identifier=data.get('identifier'),
+                version=data.get('version'),
+                date_modified=date_obj,
+                is_part_of=Project.from_json(isPartOf) if (isPartOf := data.get('isPartOf')) else None,
+                in_language=Language.from_json(inLanguage) if (inLanguage := data.get('inLanguage')) else None,
+                namespace=Namespace.from_json(namespace) if (namespace := data.get('namespace')) else None,
+                size=Size.from_json(size) if (size := data.get('size')) else None
+            )
+        except (ValueError, TypeError) as e:
+            batch_id = data.get('identifier', 'N/A')
+            raise DataModelError(f"Failed to parse Batch with identifier '{batch_id}': {e}") from e
 
     @staticmethod
     def to_json(batch: 'Batch') -> dict:
+        """Safely converts a Batch instance to a dictionary, handling optional None values."""
         return {
             'identifier': batch.identifier,
             'version': batch.version,
-            'dateModified': batch.date_modified.isoformat(),
-            'isPartOf': Project.to_json(batch.is_part_of),
-            'inLanguage': Language.to_json(batch.in_language),
-            'namespace': Namespace.to_json(batch.namespace),
-            'size': Size.to_json(batch.size)
+            'dateModified': batch.date_modified.isoformat() if batch.date_modified else None,
+            'isPartOf': Project.to_json(batch.is_part_of) if batch.is_part_of else None,
+            'inLanguage': Language.to_json(batch.in_language) if batch.in_language else None,
+            'namespace': Namespace.to_json(batch.namespace) if batch.namespace else None,
+            'size': Size.to_json(batch.size) if batch.size else None
         }
